% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkMajorizing}
\alias{checkMajorizing}
\title{checkMajorizing: Helper function to allow for various types of
majorizing functions in accrej and thinning, including user-supplied
functions or data.frame indicating piecewise constant or piecewise
linear.}
\usage{
checkMajorizing(majorizingFcn, majorizingFcnType, support)
}
\description{
checkMajorizing: Helper function to allow for various types of
majorizing functions in accrej and thinning, including user-supplied
functions or data.frame indicating piecewise constant or piecewise
linear.
}
