% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkQuants}
\alias{checkQuants}
\title{Streamlined quantile validator}
\usage{
checkQuants(q1, q2, min = 0, max = 1, minex = NA, maxex = NA, name1, name2)
}
\arguments{
\item{q1}{Minimum Quantile}

\item{q2}{Maximum Quantile}

\item{min}{Inclusive lower bound of quantiles}

\item{max}{Inclusive upper bound of quantiles}

\item{minex}{Exclusive lower bound of quantiles}

\item{maxex}{Exclusive upper bound of quantiles}

\item{name1}{Variable name for q1. If missing, retrieve via deparse}

\item{name2}{Variable name for q2. If missing, retrieve via deparse}
}
\value{
None. Terminates program with uniform stop if invalid.
}
\description{
Checks to see if a pair of quantiles is valid and that q1 < q2.
If it is not, stop the execution with a standardized error via checkVal.
}
\details{
Notice that will be q1 = q2 is invalid, as the range is 0.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
