% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{DrawPoint}
\alias{DrawPoint}
\title{Plot Text Box}
\usage{
DrawPoint(x, y, col, cex = 1.5, bg = "black")
}
\arguments{
\item{x}{X-coordinate}

\item{y}{Y-coordinate}

\item{col}{Color of point fill}

\item{cex}{Size of the point}

\item{bg}{Outline/Background of point}
}
\description{
This function plots a Point with outline
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
