% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueUtils.R
\name{GetDefaultDistroFcn}
\alias{GetDefaultDistroFcn}
\title{Default Arrivals/Service Distribution Function}
\usage{
GetDefaultDistroFcn(ptype, isArrival, numServers = 1, asList = FALSE)
}
\arguments{
\item{ptype}{Process type; can be specified by name. Default values
include "M" (arrival rate 1, service rate 10/9),
"G" (interarrivals uniform(0,2), service times
uniform(0, 1.8)), "D" (interarrivals 1.0, service times
9/10).}

\item{isArrival}{Is it an arrival or service generation}

\item{numServers}{Number of servers (for service functions)}

\item{asList}{What is the default for asList param of v* functions}
}
\description{
Returns a default distribution function
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{queueing}
\keyword{internal}
