% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueUtils.R
\name{GetDistroNotation}
\alias{GetDistroNotation}
\title{Try to determine the distribution notation ("M", "E", "G") based on the
given interarrival or service function in fcn.}
\usage{
GetDistroNotation(fcn)
}
\arguments{
\item{fcn}{The closure representing the interarrival or service
process function.}
}
\description{
Returns a distribution notation for displaying
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{queueing}
\keyword{internal}
