% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbinom.R
\name{inbinom}
\alias{inbinom}
\title{Random Variate Generation for the Negative Binomial Distribution}
\usage{
inbinom(
  u = runif(1),
  size,
  prob,
  mu,
  minPlotQuantile = 0,
  maxPlotQuantile = 0.95,
  plot = TRUE,
  showCDF = TRUE,
  showPMF = FALSE,
  showECDF = FALSE,
  show = NULL,
  maxInvPlotted = 50,
  plotDelay = 0,
  animateAll = plotDelay > 0 || plotDelay == -1,
  empColor = "red3",
  theoColor = "grey",
  showTitle = TRUE,
  respectLayout = FALSE,
  ...
)
}
\arguments{
\item{u}{Desired uniform distrubution elements in (0, 1) (default runif(1)). \cr
If NULL, show only theorhetical graphs with no generation}

\item{size}{Number of failures until experiments stop (r)}

\item{prob}{Probability of success in each trial in (0, 1)}

\item{mu}{Mean of data; alternative parameterization}

\item{minPlotQuantile}{Minimum quantile to plot}

\item{maxPlotQuantile}{Maximum quantile to plot}

\item{plot}{Whether to plot distribution}

\item{showCDF}{Whether to plot cumulative distribution (CDF)}

\item{showPMF}{Whether to plot probability density/mass (PMF)}

\item{showECDF}{Whether to plot empirical vs theorhetical CDF}

\item{show}{4: CDF, 2: PMF, 1: ECDF; sum for desired combo}

\item{maxInvPlotted}{Max number of inversions to plot before switching to quantile view}

\item{plotDelay}{Delay in seconds between CDF plots}

\item{animateAll}{Should animate all graphs (otherwise, CDF)}

\item{empColor}{Color of empirical data display}

\item{theoColor}{Color of theoretical data display}

\item{showTitle}{Whether to title the plot}

\item{respectLayout}{Should respect implemented plot setting}

\item{...}{Possible additional arguments. Currently, additional arguments not considered.}
}
\value{
A vector of Negative Binomial distributed random variates
}
\description{
Generates random variates from the Negative Binomial distribution by inversion.
Optionally graphs the population cumulative distribution function and
associated random variates, the population probability density function
and a histogram of the random variates, and the empirical cumulative
distribution function versus the population cumulative distribution function.
}
\details{
Generates random variates from the Negative Binomial distribution, and optionally,
illustrates
\itemize{
\item the use of the inverse-CDF technique,
\item the effect of random sampling variability in relation to the PMF and CDF.
}
When all of the graphics are requested,
\itemize{
\item the first graph illustrates the use of the inverse-CDF technique by
graphing the population CDF and the transformation of the random numbers
to random variates,
\item the second graph illustrates the effect of random sampling variability
by graphing the population PMF and the histogram associated with the
random variates, and
\item the third graph illustrates effect of random sampling variability by
graphing the population CDF and the empirical CDF associated with the
random variates.
}
All aspects of the random variate generation algorithm are output in red by
default, though this can be changed by specifying \code{empColor}.
All aspects of the population distribution are output in black.

The negative binomial distribution with \code{size} \eqn{= n} and
\code{prob} \eqn{= p} has density

\if{html}{\out{<div class="sourceCode">}}\preformatted{      \deqn{p(x) = \frac{\Gamma(x+n)}{\Gamma(n) \ x!} p^n (1-p)^x}{%
                p(x) = Gamma(x+n)/(Gamma(n) x!) p^n (1-p)^x}
}\if{html}{\out{</div>}}

for \eqn{x = 0, 1, 2, \ldots, n > 0} and \eqn{0 < p \leq 1}. This represents the
number of failures which occur in a sequence of Bernoulli trials before a
target number of successes is reached.

The mean is gYvY8W9z5ddmOdApoicXBuySSoAbNWVp-8\eqn{ = n(1 - p)/p} and variance \eqn{n(1 - p)/p^2}

The algorithm for generating random variates from the Negative Binomial distribution is
synchronized (one random variate for each random number) and monotone in u.
This means that the variates generated here might be useful in some variance
reduction techniques used in Monte Carlo and discrete-event simulation.

Values from the u vector are plotted in the cdf plot along the vertical axis
as colored dots.  A horizontal, dashed, colored line extends from the dot to
the population cdf.  At the intersection, a vertical, dashed colored line
extends downward to the horizontal axis, where a second colored dot, denoting
the associated gamma random variate is plotted.

This is not a particularly fast variate generation algorithm because it uses
the base R \code{qnbinom} function to invert the values contained in \code{u}.

All of the elements of the \code{u} vector must be between 0 and 1.
Alternatively, \code{u} can be \code{NULL} in which case plot(s) of the
theoretical PMF and cdf are displayed according to plotting parameter
values (defaulting to display of both the PMF and cdf).

The \code{show} parameter can be used as a shortcut way to denote plots to
display.  The argument to \code{show} can be either:
\itemize{
\item a binary vector of length three, where the entries from left to right
correspond to \code{showCDF}, \code{showPMF}, and \code{showECDF},
respectively.  For each entry, a 1 indicates the plot should be
displayed, and a 0 indicates the plot should be suppressed.
\item an integer in \link{0,7} interpreted similar to Unix's chmod command.  That
is, the integer's binary representation can be transformed into a
length-three vector discussed above (e.g., 6 corresponds to c(1,1,0)).
See examples.
}
Any valid value for \code{show} takes precedence over existing individual
values for \code{showCDF}, \code{showPMF}, and \code{showECDF}.

If \code{respectLayout} is \code{TRUE}, the function respects existing
settings for device layout.  Note, however, that if the number of plots
requested (either via \code{show} or via \code{showCDF}, \code{showPMF}, and
\code{showECDF}) exceeds the number of plots available in the current layout
(as determined by \code{prod(par("mfrow"))}), the function will display all
requested plots but will also display a warning message indicating that the
current layout does not permit simultaneous viewing of all requested plots.
The most recent plot with this attribute can be further annotated after the call.

If \code{respectLayout} is \code{FALSE}, any existing user settings for device
layout are ignored.  That is, the function uses \code{par} to explicitly set
\code{mfrow} sufficient to show all requested plots stacked vertically to
align their horizontal axes, and then resets row, column, and margin settings
to their prior state on exit.

The \code{minPlotQuantile} and \code{maxPlotQuantile} arguments are present in
order to compress the plots horizontally.   The random variates generated are
not impacted by these two arguments.  Vertical, dotted, black lines are
plotted at the associated quantiles on the plots.

\code{plotDelay} can be used to slow down or halt the variate generation for
classroom explanation.  The \code{animateAll} option (enabled by default if
plotDelay > 0 or -1) option will generate all of the plots at the same time
with only consideration of the subset of data that has been inverted so far.
In the plot associated with the PMF, the maximum plotting height is
associated with 125\\% of the maximum height of PMF. Any histogram cell
that extends above this limit will have three black dots appearing above it.
}
\examples{
 inbinom(0.5, size = 10, mu = 10)

 set.seed(8675309)
 inbinom(runif(10), 10, 0.25, showPMF = TRUE)

 set.seed(8675309)
 inbinom(runif(10), 10, 0.25, showECDF = TRUE)

 set.seed(8675309)
 inbinom(runif(10), 10, 0.25, showPMF = TRUE, showECDF = TRUE, empColor = "blue3")

 set.seed(8675309)
 inbinom(runif(10), 10, 0.25, showPMF = TRUE, showCDF = FALSE)

 inbinom(runif(100), 10, 0.25, showPMF = TRUE, minPlotQuantile = 0.02, maxPlotQuantile = 0.98)

 # plot the PMF and CDF without any variates
 inbinom(NULL, 10, 0.25, showPMF = TRUE, showCDF = TRUE)

 # plot CDF with inversion and PMF using show
 inbinom(runif(10), 10, 0.25, show = c(1,1,0))
 inbinom(runif(10), 10, 0.25, show = 6)

 # plot CDF with inversion and ECDF using show
 inbinom(runif(10), 10, 0.25, show = c(1,0,1))
 inbinom(runif(10), 10, 0.25, show = 5)

 # plot CDF with inversion, PMF, and ECDF using show
 inbinom(runif(10), 10, 0.25, show = c(1,1,1))
 inbinom(runif(10), 10, 0.25, show = 7)

 # plot three different CDF+PMF+ECDF vertical displays
 par(mfcol = c(3,3))  # 3 rows, 3 cols, filling columns before rows
 set.seed(8675309)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE)

 # plot three different CDF+PMF+ECDF horizontal displays,
 # with title only on the first display
 par(mfrow = c(3,3))  # 3 rows, 3 cols, filling rows before columns
 set.seed(8675309)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE, showTitle = FALSE)
 inbinom(runif(20), 10, 0.25, show = 7, respectLayout = TRUE, showTitle = FALSE)

 # exhibit use of the respectLayout = FALSE option (default)
 par(mfrow = c(3,3))  # this will be ignored below since respectLayout = FALSE
 par(mar = c(5.1, 4.1, 4.1, 2.1))
 set.seed(8675309)
 inbinom(runif(20), 10, 0.25, show = 7)
 par("mfrow")  # will have been reset to c(3,3)
 par("mar")    # will have been reset to c(5.1, 4.1, 4.1, 2.1)


 # overlay visual exploration of ks.test results
 set.seed(54321)
 vals <- inbinom (runif(10), 10, 0.25, showECDF = TRUE)
 D <- as.numeric(ks.test(vals, "pnbinom", 10, 0.25)$statistic)
 for (x in seq(0.75, 1.25, by = 0.05)) {
  y <- pnbinom(x, 10, 0.25)
  segments(x, y, x, y + D, col = "darkgreen", lwd = 2, xpd = NA)
 }

 # allow animation of CDF component only
 inbinom(runif(10), 10, 0.25, show = c(1,1,0), plotDelay = 0.1, animateAll = FALSE)
 inbinom(runif(10), 10, 0.25, show =        7, plotDelay = 0.1, animateAll = FALSE)

 # allow animation of all component
 inbinom(runif(10), 10, 0.25, show = c(1,1,0), plotDelay = 0.1, animateAll = TRUE)
 inbinom(runif(10), 10, 0.25, show =        7, plotDelay = 0.1)

 # pause at each stage of inversion
 inbinom(runif(10), 10, 0.25, show =        7, plotDelay = -1)

}
\seealso{
\code{\link[=runif]{stats::runif}}

\code{\link[=rnbinom]{stats::rnbinom}}
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{random variate generation}
\keyword{distribution}
\keyword{dynamic}
\keyword{hplot}
