% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.color}
\alias{is.color}
\title{Color validator}
\usage{
is.color(c)
}
\arguments{
\item{c}{The color in question}
}
\value{
True if parsing the input as a color will pass.
}
\description{
Check to see if the input can be parsed into a color
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
