% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isValNum}
\alias{isValNum}
\title{Checks to see if a variable is a valid number of vector of numbers}
\usage{
isValNum(n, l = 1)
}
\arguments{
\item{n}{The number in question}

\item{l}{valid lengths of the vector (can be vector)}
}
\value{
Logical
}
\description{
Checks to see if a variable is a valid number of vector of numbers
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
