% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehmer.R
\name{lehmer}
\alias{lehmer}
\title{Lehmer Generator Simulation}
\usage{
lehmer(
  a = 13,
  m = 31,
  seed = 1,
  maxSteps = Inf,
  showOutput = TRUE,
  showTitles = TRUE,
  plotDelay = -1,
  sf_ratio = c(3, 1)
)
}
\arguments{
\item{a}{Multiplier in equation}

\item{m}{Modulus component in equation}

\item{seed}{Seed in equation; the initial X value x_0}

\item{maxSteps}{Maximum number of steps to display; infinite by default
Ignored if showOutput is false}

\item{showOutput}{Should the visual output be displayed, if not, it will just return output}

\item{showTitles}{Should the title be displayed in the main plot}

\item{plotDelay}{Wait time between transitioning; -1 (default) for interactive
mode, where the user is queried for input to progress}

\item{sf_ratio}{Multiplied for font scaling ratio relative to screen x/y ratio.
Set to c(2, 1) by default with ideal screening at a 2:1 ratio}
}
\value{
The PRNG cycle. If multiple possible options, will do one of the following. The path returned
is the last path encountered by the generator in its visual generation loop. If showOutput is
FALSE, it will return the first found one.
}
\description{
This function animates the processes of a basic Lehmer
pseudo-random number generator. Also known in the literature as a
multiplicative linear congruential generator (MLCG), the generator
operates on the formula:

\deqn{X_{k+1} \equiv a \cdot X_k \pmod{m}}{%
      X_{k+1} = a * X_k mod m}
}
\examples{
 # Default case (m, a = 31, 13); small full period
 lehmer(maxStep = 50)                   # Interactive mode (plotDelay = -1)
 lehmer(plotDelay = 0.1, maxStep = 50)  # Auto-progressing
 
 # Change behavior of fonts relative to x/y ratio of display device
 #Should be customized per device if desired
 lehmer(plotDelay = 0.1, maxStep = 50, sf_ratio=c(3, 1))  # Good for wider display
 lehmer(plotDelay = 0.1, maxStep = 50, sf_ratio=c(1, 3))  # Good for taller display

 # Special case with booleanic generator
 lehmer(a = 11, m = 20, maxStep = 50, plotDelay = 0.1)

 # Special edge case where the seed changes because 1x13 == (78x13) mod 1001 
 lehmer(a = 12, m = 20, maxStep = 50, sf_ratio=c(1, 3), plotDelay = 0.1)
 
}
\concept{random variate generation}
