% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueUtils.R
\name{ParseDistroFcn}
\alias{ParseDistroFcn}
\title{If the provided interarrival or service function does not allow for
an arbitrary number of arguments (using ...), which is required by the
simEd animation code, rebuild and return that function to have '...' as
its last parameter.}
\usage{
ParseDistroFcn(fcn, asList = FALSE)
}
\arguments{
\item{fcn}{The rebuilt closure representing the interarrival or service
process function.}
}
\description{
Rebuilds a function to include '...' parameter, as necessary.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{queueing}
\keyword{internal}
