% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ParseShow}
\alias{ParseShow}
\title{Distributes a parameter into components}
\usage{
ParseShow(showBools = FALSE, show = NULL, ignoreBools = FALSE)
}
\arguments{
\item{showBools}{Vector of logical values (booleans)}

\item{show}{A value to be distributed among the booleans.
Can be an empty value (i.e. NULL, NA, c()), a logical of
length 1 or \code{length(showBools)}, or a single integer
to be parsed a la Unix's chmod. (see 'Values' for details)}

\item{ignoreBools}{If TRUE, disregard the showBools values while processing.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Updated versions of showBools. Specifically:
\\itemize\{
   \\item If show is an empty value (i.e. NULL, NA, c()), return the
       inputted showBools unaltered
   \\item If show is a 1-length logical (TRUE/FALSE), apply it to all of
       showBools and return them.
   \\item If show is a vector of elements with same length as showBools,
       distribute it among showBools and return the result.
   \\item If show is a number, parse it a la Unix's chmod and distribute
       For example, 7 -> (4 + 2 + 1) -> c(TRUE, TRUE, TRUE),
       6 -> (4 + 2 + 0) -> c(TRUE, TRUE, FALSE)
 \}
}\if{html}{\out{</div>}}
}
\description{
Given a parameter show and a set of parameters showBools
which are related, parse the show parameter in one of four ways
(see Values). This allows the user to specify T/F values either
vurbosely or concisely.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
