% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{PausePlot}
\alias{PausePlot}
\title{Pause Plot For User Input}
\usage{
PausePlot(
  pauseData,
  prompt = NA,
  viewCommand = c(),
  viewInstruct = c(),
  viewFunction = c(),
  currStep = 0
)
}
\arguments{
\item{prompt}{Default prompt for PausePlot calls}

\item{viewCommand}{A vector of strings a user can input to call custom
view functions}

\item{viewInstruct}{A vector of string instructions to display alongside
default view instructions. Should correspond with
viewCommand values.}

\item{viewFunction}{A list of functions to be called when a commmand is
issued via prompt. Should correspond with viewCommand.
Must have at least one parameter, which should expect
the value inputted by user after command string.}

\item{currStep}{A numerical value representing a 'current step' count.
Specifying this will enable jump command. Should be positive}

\item{plotDelay}{Current delay setting, in secs. If -1, prompt for input}
}
\value{
An updated plotDelay
}
\description{
Handles pausing of plot at regular intervals depending on plotDelay.
Also handles user input and processing for plotDelay == -1
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
