% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotContinuous.R
\name{PlotContinuous}
\alias{PlotContinuous}
\title{i* Plotting Function for Continuous Distributions}
\usage{
PlotContinuous(
  u = runif(1),
  minPlotQuantile = 0.05,
  maxPlotQuantile = 0.95,
  plot = TRUE,
  showCDF = TRUE,
  showPDF = FALSE,
  showECDF = FALSE,
  show = NULL,
  maxInvPlotted = 50,
  plotDelay = 0,
  animateAll = TRUE,
  empColor = "red3",
  theoColor = "grey",
  showTitle = TRUE,
  respectLayout = FALSE,
  getDensity,
  getDistro,
  getQuantile,
  hasCDF,
  hasPDF,
  hasECDF,
  titleStr = ""
)
}
\arguments{
\item{u}{Desired uniform distrubution elements in (0, 1) (default runif(1)). \cr
If NULL, show only theorhetical graphs with no generation}

\item{minPlotQuantile}{Minimum quantile to plot}

\item{maxPlotQuantile}{Maximum quantile to plot}

\item{plot}{Whether to plot distribution}

\item{showCDF}{Whether to plot cumulative distribution (CDF)}

\item{showPDF}{Whether to plot probability density (PDF)}

\item{showECDF}{Whether to plot empirical vs theorhetical CDF}

\item{show}{4: CDF, 2: PDF, 1: ECDF; sum for desired combo}

\item{maxInvPlotted}{Max number of inversions to plot before switching to quantile view}

\item{plotDelay}{Delay in seconds between plots. If -1, use interactive mode}

\item{animateAll}{Should animate all graphs (otherwise, CDF)}

\item{empColor}{Color of empirical data display}

\item{theoColor}{Color of theorhetical data display}

\item{showTitle}{Whether to title the plot}

\item{respectLayout}{Should respect implemented plot setting}

\item{getDensity}{A density function for the distribution (i.e. \code{dunif} for uniform)
\cr (REQUIRED w/ NO DEFAULTS)}

\item{getDistro}{A distribution function for the distribution (i.e. \code{punif} for uniform)
\cr (REQUIRED w/ NO DEFAULTS)}

\item{getQuantile}{A quantile function for the distribution (i.e. \code{qunif} for uniform)
\cr (REQUIRED w/ NO DEFAULTS)}

\item{hasCDF}{Tells function if \code{showCDF} was specified. Used for determining
priority of individual show- parameters and the main show parameter}

\item{hasPDF}{Tells function if \code{showPDF} was specified. Used for determining
priority of individual show- parameters and the main show parameter}

\item{hasECDF}{Tells function if \code{showECDF} was specified. Used for determining
priority of individual show- parameters and the main show parameter}

\item{titleStr}{String/Language of text to be displayed as the title}
}
\value{
A vector of random variates distributed according to the provided
distribution function
}
\description{
Performs displays for i* functions (i.e. \code{iexp}), in which the parameters
are leveraged to plot the CDF, PDF, and ECDF of the distribution.
Used internally in continuous i* functions, but can be used externally
to define other distributions with definable density, distribution, and
quantile functions.
}
\details{
Generates random variates using the inputted getDistro function and, optionally,
illustrates
\itemize{
\item the use of the inverse-CDF technique,
\item the effect of random sampling variability in relation to the <\%= PXF \%> and CDF.
}
When all of the graphics are requested,
\itemize{
\item the first graph illustrates the use of the inverse-CDF technique by
graphing the population CDF and the transformation of the random numbers
to random variates,
\item the second graph illustrates the effect of random sampling variability
by graphing the population <\%= PXF \%> and the histogram associated with the
random variates, and
\item the third graph illustrates effect of random sampling variability by
graphing the population CDF and the empirical CDF associated with the
random variates.
}
The functionality of this function is a generalization of the functionality
continuous i* functions such as \code{ibeta}, which should be considered for
more detail on these parameters. Plotting functionality for i* functions is
provided by \code{PlotContinuous} as well as its discrete distribution
counterpart, \code{PlotDiscrete}.
}
\seealso{
\code{\link[=runif]{stats::runif}}
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{random variate generation}
