% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{PlotTimer}
\alias{PlotTimer}
\title{Times Execution and Plots Results}
\usage{
PlotTimer(
  times = c(),
  plotRange = c(160, 200, 0, 30),
  borderDisp = c(-15, 0, 0, 0),
  start = TRUE,
  print = FALSE,
  showPlot = FALSE
)
}
\arguments{
\item{times}{Times component. Generated on initial PlotTimer call.
Last element is the most recent time record or change in
time recorded; all previous entries are previous changes in
time, in order of recording time.}

\item{plotRange}{Coordinate range of output. Consistent with TogglePlot specs.
4-element vector with x0, x1, y0, and y1, in that order.}

\item{borderDisp}{Displacement component used for moving background border.
4-element vector with x0, x1, y0, and y1, in that order.}

\item{start}{Whether or not the timer should start with this call.
If yes, last element will be set to System time (last recorded time)
If no, last element will be set to 0 (latest change in time)}
}
\value{
Updated vector that should be inputted into subsequent PlotTimer/ToggleTimer calls.
Last element is either last-recorded time or change in time. Prior elements are
the changes in time recorded thus far.
}
\description{
Helper function to help visualize component timing. Maintains running
times list which is returned and should be passed in. Also serves
as a TogglePlot use-case.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
