% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pround}
\alias{pround}
\title{Uniform rounder for printing}
\usage{
pround(n)
}
\arguments{
\item{n}{Numerical value or a vector of such}
}
\value{
A formatted string of vector of strings such that, decimals are
rounded to exactly 3 decimal places, integers are returned unrounded, and
NaN values are replaced with "-"
}
\description{
This function rounds all values passed in to a specific decimal
point and formats it for printing
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
