% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resize}
\alias{resize}
\title{Vector Resizer}
\usage{
resize(vec)
}
\arguments{
\item{vec}{A vector to be resized}
}
\value{
A resized vector padded with NaNs.
}
\description{
Given a vector of length n, this will return a vector of length 2n where
the first 1:n entries from the original are copied into the new vector.
Using this avoids the memory hits of always using the c() function.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
