% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{ScaleFont}
\alias{ScaleFont}
\title{Font-Scaling Function}
\usage{
ScaleFont(n, f = 100, r = c(1, 1))
}
\arguments{
\item{n}{Size to scale to}
}
\value{
A numeric value of the scaled n
}
\description{
Returns a scaling of the inputted number based on the device.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
