% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{SetPausePlot}
\alias{SetPausePlot}
\title{Pause Plot Value Setter}
\usage{
SetPausePlot(
  plotDelay = -1,
  prompt = NA,
  viewCommand = c(),
  viewInstruct = c(),
  viewFunction = c(),
  recLimit = 5
)
}
\arguments{
\item{prompt}{Default prompt for PausePlot calls}

\item{viewCommand}{Default view command for PausePlot calls}

\item{viewInstruct}{Default view instruction for PausePlot calls}

\item{viewFunction}{Default view function for PausePlot calls}
}
\description{
Overrides currently set/default value of PausePlot calls.
}
\details{
See PausePlot for more thorough description of parameters
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
