% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssqvis.R
\name{ssqvis}
\alias{ssqvis}
\title{Single-Server Queue Simulation Visualization}
\usage{
ssqvis(
  maxArrivals = Inf,
  seed = NA,
  interarrivalType = "M",
  serviceType = "M",
  maxTime = Inf,
  maxDepartures = Inf,
  maxInSystem = Inf,
  maxEventsPerSkyline = 15,
  saveAllStats = FALSE,
  saveInterarrivalTimes = FALSE,
  saveServiceTimes = FALSE,
  saveWaitTimes = FALSE,
  saveSojournTimes = FALSE,
  saveNumInQueue = FALSE,
  saveNumInSystem = FALSE,
  saveServerStatus = FALSE,
  showOutput = TRUE,
  showSkylineInQueue = TRUE,
  showSkylineInSystem = TRUE,
  showSkylineInServer = TRUE,
  showSkyline = NULL,
  showTitles = TRUE,
  showProgress = FALSE,
  jobImage = NA,
  plotDelay = -1
)
}
\arguments{
\item{maxArrivals}{maximum number of customer arrivals allowed to enter the system}

\item{seed}{initial seed to the random number generator (NA uses current state of
random number generator; NULL seeds using system clock)}

\item{interarrivalType}{string representation of desired interarrival process
(default is "M": exponential with rate 1).}

\item{serviceType}{string representation of desired service process
(default is "M": exponential with rate 10/9).}

\item{maxTime}{maximum time to simulate}

\item{maxDepartures}{maximum number of customer departures to process}

\item{maxInSystem}{maximum number of customers that the system can hold (server(s) plus queue).
Infinite by default.}

\item{maxEventsPerSkyline}{maximum number of events viewable at a time in the skyline plot.
A large value for this parameter may result in plotting delays.
This parameter does not impact the final plotting, which will show all
end-of-simulation results.}

\item{saveAllStats}{if TRUE, returns all vectors of statistics (see below)
collected by the simulation}

\item{saveInterarrivalTimes}{if TRUE, returns a vector of all interarrival times generated}

\item{saveServiceTimes}{if TRUE, returns a vector of all service times generated}

\item{saveWaitTimes}{if TRUE, returns a vector of all wait times (in the queue) generated}

\item{saveSojournTimes}{if TRUE, returns a vector of all sojourn (time in the system)
times generated}

\item{saveNumInQueue}{if TRUE, returns a vector of times and a vector of counts for whenever
the number in the queue changes}

\item{saveNumInSystem}{if TRUE, returns a vector of times and a vector of counts for whenever
the number in the system changes}

\item{saveServerStatus}{if TRUE, returns a vector of times and a vector of
server status (0:idle, 1:busy) for whenever the status changes}

\item{showOutput}{if TRUE, displays summary statistics upon completion}

\item{showSkylineInQueue}{if TRUE, the skyline will track the number of elements in the queue.}

\item{showSkylineInSystem}{if TRUE, the skyline will track the number of elements in the system.}

\item{showSkylineInServer}{if TRUE, the skyline will track the number of elements in the server.}

\item{showSkyline}{Shorthand for specifying showSkyline... parameters using chmod-like
octal component specification: use 1, 2, 4 for system, queue, and server
respectively, summing to indicate desired combination (e.g., 7 for all).}

\item{showTitles}{if TRUE, titles will be shown on plots}

\item{showProgress}{if TRUE, displays a progress bar on screen during execution}

\item{jobImage}{a vector of URLs/local addresses of images to use as jobs. Requires
package 'Magick'.}

\item{plotDelay}{a positive numeric value indicating seconds between state plottings.
A value of -1 enters 'interactive' mode, where the state will pause
for user input at each step.  A value of 0 will display only the final
end-of-simulation plot.}
}
\value{
The function returns a list containing:
\itemize{
\item the number of arrivals to the system (\code{customerArrivals}),
\item the number of customers processed (\code{customerDepartures}),
\item the ending time of the simulation (\code{simulationEndTime}),
\item average wait time in the queue (\code{avgWait}),
\item average time in the system (\code{avgSojourn}),
\item average number in the system (\code{avgNumInSystem}),
\item average number in the queue (\code{avgNumInQueue}), and
\item server utilization (\code{utilization}).
}
of the queue as computed by the simulation.
When requested via the ``save...'' parameters, the list may also contain:
\itemize{
\item a vector of interarrival times (\code{interarrivalTimes}),
\item a vector of wait times (\code{waitTimes}),
\item a vector of service times (\code{serviceTimes}),
\item a vector of sojourn times (\code{sojournTimes}),
\item two vectors (time and count) noting changes to number in the system
(\code{numInSystemT}, \code{numInSystemN}),
\item two vectors (time and count) noting changes to number in the queue
(\code{numInQueueT}, \code{numInQueueN}), and
\item two vectors (time and status) noting changes to server status
(\code{serverStatusT}, \code{serverStatusN}).
}
}
\description{
A modified ssq implementation that illustrates event-driven
details, including the event calendar, inversion for interarrival and service
time variate generation, the simulation clock, the status of the queueing
system, and statistics collection, plotting step-by-step in either an
interactive mode or time-delayed automatic mode.
}
\details{
Illustrates and animates the details of an event-driven implementation of a
single-server queue simulation.

Animates the details of an event-driven implementation of a single-server
queue simulation.

The event calendar, inversion for interarrival and service time variates,
and an abbreviated (current) timeline are animated in the top pane of the
window.  In this pane, blue corresponds to the arrival process, orange
corresponds to the service process, and purple corresponds to uniform
variates used in inversion.  Yellow is used to highlight recent updates.

The state of the queueing system is animated in the middle pane of the
window.  In this pane, red indicates an idle server, orange indicates that
a new customer has just arrived to the server and a corresponding service
time is being generated, and green indicates a busy server.  By default,
customers are depicted as black rectangles and identified by increasing
arrival number, but this depiction can be overridden by the \code{jobImage}
parameter.

Statistics are displayed in the bottom pane of the window.  Time-persistent
statistics are shown as "skyline functions" in the left portion of this
pane.  Both time-persistent and based-on-observation statistics are shown
in respective tables in the right portion of this pane. In the tables,
yellow is used to highlight recent updates.

The \code{seed} parameter can take one of three valid
argument types:
\itemize{
\item \code{NA} (default), which will use the current state of the random
number generator without explicitly setting a new seed (see examples);
\item a positive integer, which will be used as the initial seed passed in
an explicit call to \code{\link{set.seed}}; or
\item \code{NULL}, which will be passed in an explicit call to to
\code{\link{set.seed}}, thereby setting the initial seed using the
system clock.
}
}
\examples{

 # Visualizing ssq with a set seed, infinite queue capacity, 20 arrivals,
 # and showing skyline for all 3 attributes
 ssqvis(seed = 1234, maxArrivals = 20, showSkyline = 7)

 # Perform simulation again with finite queue of low capacity. Note same
 # variable generation but different outcomes due to rejection pathway
 ssqvis(seed = 1234, maxArrivals = 25, showSkyline = 7, maxInSystem = 5)

 # Perform simulation again with finite queue of low capacity. Note same
 # variable generation but different outcomes due to rejection pathway
 ssqvis(seed = 1234, maxTime = 25, showSkyline = 7, maxInSystem = 5)

 # Using default distributions to make a default M/G/1 Queue
 ssqvis(seed = 1234, maxDepartures = 10, interarrivalType = "M", serviceType = "G")

}
\seealso{
\code{\link{rstream}}, \code{\link{set.seed}},
\code{\link[=runif]{stats::runif}}
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\concept{queueing}
\keyword{utilities}
