% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{TextBox}
\alias{TextBox}
\title{Plot Text Box}
\usage{
TextBox(
  text = "",
  mw,
  mh,
  hw = NA,
  hh = hw,
  col = "black",
  bg = NA,
  border = "black",
  img = NA,
  font = 1,
  size = 15,
  txd = 0,
  tyd = 0,
  textf = 0,
  xs = 0,
  ys = 0
)
}
\arguments{
\item{text}{Text to be displayed in the box}

\item{mw}{Mid-width of the textbox}

\item{mh}{Mid-height of the textbox}

\item{hw}{Half-width of the textbox}

\item{hh}{Half-height of the textbox}

\item{col}{Color of text}

\item{bg}{Color of background box (NA for no background)}

\item{img}{Raster image to be desplayed as background}

\item{font}{Font style of text
(1 = plain, 2 = bold, 3 = italic, 4 = bold-italic)}

\item{size}{Scaled font-size of text}

\item{txd}{Text x-displacement from center as a percentage of half-width}

\item{tyd}{Text y-displacement from center as a percentage of half-height}

\item{textf}{Function to be applied to text prior to displaying}

\item{xs}{Scaling function to be applied to x}

\item{ys}{Scaling function to be applied to y}
}
\description{
This function plots a TextBox with a variety of options including
a scaled raster image for the background
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
