% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thinning}
\alias{thinning}
\title{Thinning}
\usage{
thinning(
  maxTime = 24,
  intensityFcn = function(x) (5 - sin(x/0.955) - (4 * cos(x/3.82)))/0.5,
  majorizingFcn = NULL,
  majorizingFcnType = NULL,
  seed = NA,
  maxTrials = Inf,
  plot = TRUE,
  showTitles = TRUE,
  plotDelay = plot * -1,
  sf_ratio = c(1, 3)
)
}
\arguments{
\item{maxTime}{Maximum time of the generated non-homogeneous Poisson process.}

\item{intensityFcn}{Intensity function corresponding to rate of arrivals across
time.}

\item{majorizingFcn}{Majorizing function.  Default value is NULL,
corresponding to a constant majorizing function.  May
alternatively be provided as a user-specified function,
or as a data frame requiring additional notation as
either piecewise-constant or piecewise-linear.  See
examples.}

\item{majorizingFcnType}{Used to indicate whether a majorizing function that
is provided via data frame is to be interpreted as
either piecewise-constant ("pwc") or piecewise-linear
("pwl").  If the majorizing function is either the
default or a user-specified function (closure), the
value of this parameter is ignored.}

\item{seed}{Initial seed for the uniform variates used during
generation.}

\item{maxTrials}{Maximum number of accept-reject trials; infinite by default.}

\item{plot}{If TRUE, visual display will be produced.  If FALSE,
generated variates will returned without visual display.}

\item{showTitles}{Should the title be displayed in the main plot}

\item{plotDelay}{Wait time in seconds between plots; -1 (default) for
interactive mode, where the user is queried for input
to progress.}

\item{sf_ratio}{Multiplied for font scaling ratio relative to screen x/y ratio.
Set to c(2, 1) by default with ideal screening at a 2:1 ratio}
}
\value{
Returns the generated variates pulled from the actual distribution
}
\description{
This function animates the generation of a non-homogeneous
Poisson process via the "thinning" approach, given a specified
intensity function and majorizing function.
}
\examples{

nhpp <- thinning()
nhpp <- thinning(maxTime = 48, seed = 8675309, plotDelay = 0)
nhpp <- thinning(maxTime = 8, seed = 8675309, plotDelay = 0.01)

# piecewise-constant data.frame for bounding default intensity function
fpwc <- data.frame(
    x = c(0, 2, 20, 30, 44, 48),
    y = c(5, 5, 20, 12, 20,  5)
)
nhpp <- thinning(maxTime = 24, plotDelay = 0, 
                 majorizingFcn = fpwc, majorizingFcnType = "pwc")

# piecewise-linear data.frame for bounding default intensity function
fpwl <- data.frame(
    x = c(0, 12, 24, 36, 48),
    y = c(5, 25, 10, 25, 5)
)
nhpp <- thinning(maxTime = 24, plotDelay = 0, 
                 majorizingFcn = fpwl, majorizingFcnType = "pwl")

# piecewise-linear closure/function bounding default intensity function
fclo <- function(x) {
    if (x <= 12) (5/3)*x + 5
    else if (x <= 24) 40 - (5/4)*x
    else if (x <= 36) (5/4)*x - 20
    else 85 - (5/3) * x
}
nhpp <- thinning(maxTime = 48, plotDelay = 0, majorizingFcn = fclo)

# thinning with custom intensity function and default majorizing
intensity <- function(x) { 
    day <- 24 * floor(x/24)
    return(80 * (dnorm(x, day + 6,    2.5) + 
                 dnorm(x, day + 12.5, 1.5) + 
                 dnorm(x, day + 19,   2.0)))
}
nhpp <- thinning(maxTime = 24, plotDelay = 0, intensityFcn = intensity)

# piecewise-linear data.frame for bounding custom intensity function
fpwl <- data.frame(
    x = c(0,  6,  9, 12, 16, 19, 24, 30, 33, 36, 40, 43, 48),
    y = c(5, 17, 12, 28, 14, 18,  7, 17, 12, 28, 14, 18,  7)
)
nhpp <- thinning(maxTime = 48, plotDelay = 0, intensityFcn = intensity,
          majorizingFcn = fpwl, majorizingFcnType = "pwl")

}
\concept{random variate generation}
