% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.R
\name{ToggleTimer}
\alias{ToggleTimer}
\title{Toggles Timer State To Either Stop or Start Recording}
\usage{
ToggleTimer(times = c())
}
\arguments{
\item{times}{Times component. Generated on initial PlotTimer call.
Last element is the most recent time record or change in
time recorded; all previous entries are previous changes in
time, in order of recording time.}
}
\value{
Updated vector with updated state
}
\description{
Helper function to help visualize component timing. Toggles tail element of
PlotTimer output vector betweeen change in time and actual time. In effect,
tells timer to pause or resume timing.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
