% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRuns.R
\name{UnitTest}
\alias{UnitTest}
\title{Unit Testing Function}
\usage{
UnitTest(str)
}
\arguments{
\item{str}{A string to show which function to run a test on (i.e. "ibeta")}
}
\description{
Allows developers to visually diagnose issues with various i* functions.
Executions will go through a series of unit cases and prompt the user to
hit ENTER enter to continue. Unit tests are not all supposed to run, and
intentionally throw illegal values to show error checking. Currently,
this is implemented for i* functions only.
}
\author{
Barry Lawson (\email{blawson@richmond.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vadim.kudlay@richmond.edu})
}
\keyword{internal}
