% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{vgeom}
\alias{vgeom}
\title{Variate Generation for Geometric Distribution}
\usage{
vgeom(n, prob, stream = NULL, antithetic = FALSE, asList = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{prob}{Probability of success in each trial (0 \eqn{<} \code{prob} \eqn{\le} 1)}

\item{stream}{if \code{NULL} (default), uses \code{\link[=runif]{stats::runif}}
to generate uniform variates to invert via
\code{\link[=qgeom]{stats::qgeom}};
otherwise, an integer in 1:25 indicates the \code{\link[rstream]{rstream}} stream
from which to generate uniform variates to invert via
\code{\link[=qgeom]{stats::qgeom}};}

\item{antithetic}{if \code{FALSE} (default), inverts \eqn{u} = uniform(0,1)
variate(s) generated via either \code{\link[=runif]{stats::runif}} or
\code{\link[rstream]{rstream.sample}}; otherwise, uses
\eqn{1 - u}}

\item{asList}{if \code{FALSE} (default), output only the generated
random variates; otherwise, return a list with components suitable for
visualizing inversion. See return for details}
}
\value{
If \code{asList} is FALSE (default), return a vector of random variates.

Otherwise, return a list with components suitable for visualizing inversion,
specifically:
\item{u}{        A vector of generated U(0,1) variates }
\item{x}{        A vector of geometric random variates }
\item{quantile}{ Parameterized quantile function }
\item{text}{     Parameterized title of distribution }
}
\description{
Variate Generation for Geometric Distribution
}
\details{
Generates random variates from the geometric distribution.

Geometric variates are generated by inverting uniform(0,1) variates
produced either by \code{\link[=runif]{stats::runif}} (if \code{stream} is
\code{NULL}) or by \code{\link[rstream]{rstream.sample}}
(if \code{stream} is not \code{NULL}).
In either case, \code{\link[=qgeom]{stats::qgeom}} is used to
invert the uniform(0,1) variate(s).
In this way, using \code{vgeom} provides a monotone and synchronized
binomial variate generator, although not particularly fast.

The stream indicated must be an integer between 1 and 25 inclusive.

The geometric distribution with parameter \code{prob} = \eqn{p} has density
\deqn{p(x) = p (1-p)^x}
for \eqn{x = 0, 1, 2, \ldots}, where \eqn{0 < p \le 1}.
}
\examples{
 set.seed(8675309)
 # NOTE: following inverts rstream::rstream.sample using stats::qgeom
 vgeom(3, prob = 0.3)

 set.seed(8675309)
 # NOTE: following inverts rstream::rstream.sample using stats::qgeom
 vgeom(3, 0.3, stream = 1)
 vgeom(3, 0.3, stream = 2)

 set.seed(8675309)
 # NOTE: following inverts rstream::rstream.sample using stats::qgeom
 vgeom(1, 0.3, stream = 1)
 vgeom(1, 0.3, stream = 2)
 vgeom(1, 0.3, stream = 1)
 vgeom(1, 0.3, stream = 2)
 vgeom(1, 0.3, stream = 1)
 vgeom(1, 0.3, stream = 2)

 set.seed(8675309)
 variates <- vgeom(100, 0.3, stream = 1)
 set.seed(8675309)
 variates <- vgeom(100, 0.3, stream = 1, antithetic = TRUE)

}
\seealso{
\code{\link[rstream]{rstream}}, \code{\link{set.seed}},
\code{\link[=runif]{stats::runif}}

\code{\link[=rgeom]{stats::rgeom}}
}
\author{
Barry Lawson (\email{blawson@bates.edu}), \cr
Larry Leemis (\email{leemis@math.wm.edu}), \cr
Vadim Kudlay (\email{vkudlay@nvidia.com})
}
\concept{random variate generation}
\keyword{distribution}
